+++ Tests/XcodeGenKitTests/ProjectSpecTests.swift
@@ -1,6 +1,7 @@
 import Spectre
 import XcodeGenKit
 import xcproj
+import PathKit
 import ProjectSpec
 
 func projectSpecTests() {
@@ -99,6 +100,14 @@ func projectSpecTests() {
                 try expectValidationError(spec, .invalidFileGroup("invalidFileGroup"))
                 try expectValidationError(spec, .invalidSettingsGroup("invalidSettingGroupSettingGroup"))
                 try expectValidationError(spec, .invalidBuildSettingConfig("invalidSettingGroupConfig"))
+            }
+
+            $0.it("allows non-existent configurations") {
+                var spec = baseSpec
+                spec.options = ProjectSpec.Options(disabledValidations: [.missingConfigs])
+                let configPath = fixturePath + "test.xcconfig"
+                spec.configFiles = ["missingConfiguration": configPath.string]
+                try spec.validate()
             }
 
             $0.it("fails with invalid target") {